requirements.txt
Flask
aiohttp
bs4
pymongo
python-dotenv


